mgr/balancer: optimized info level logs
Improved the quality of one of the info level logs by
adding more information to it and removing a
similar log in debug level. This is in an effort to
provide maximum information at info level without
clogging up the info level logs with multiple logs.
I've also added a check for the "balancing root"
logs which are repeated multiple times without
adding new information. Now only the first of the
repeating "balancing root" logs is logged at info
level, the rest will be logged in debug level.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>